[C++ API] Step 3: add the new tensor API files
This PR adds the files refactoring the tensor API.

In particular this:
1. adds a DLConstTensor for proper const correctness propagation + proper
overloads and specializations
2. uniformizes tensor related functions and puts them in the same location
3. makes TensorInfo more visible as the tensor representation on which
compilation and cache interaction occur
4. cleanup unused functions
5. makes this file the only place that includes <dlpack/dlpack.h>

As part of the bigger refactoring, these files are not yet activated
(or even compiled) therefore the changeset cannot be tested independently.
This new implementation will be switched in as part of the last commit of the
the global refactoring.