[C++ API] Step 6: add the new OptionsCache
This PR adds the files implementing the OptionsCache API (#339).
Options cache is used only in the autotuner as a consequence it is moved in the
proper directory. Upon further inspection and rewriting of the autotuner logic
I found the cache implementation was convoluted by virtue of trying to do
too much factoring out with the CRTP-based cache used for compilation.

This is a simpler implementation that uses the same key/value representation
in both proto and memory (for collision avoidance, the proto is not submitted
in this PR and will land at the end).
A lightweight wrapper with conversion to/from proto is provided as usual because
the proto API doesn't hold a candle to STL.
The implementation is currently using ```std::unordered_multimap``` there are
some inefficiencies related to linear search and duplication of keys in the
proto. These can go away after refactoring #307 is completed along the lines of:
1. switch to proto3 which has support for map
2. do better than linear search

The above should be backed by data: case 1 will likely happen when we start
having larger options files but the benefits should be a factor 2 at most
(see comment in proto in #339); case 2 should not be a problem unless we start
storing much more than 10 options per TC. In any case I do not see anything
suspicious when running the tuner locally.

As part of the bigger refactoring, these files are not yet activated
(or even compiled) therefore the changeset cannot be tested independently.
This new implementation will be switched in as part of the last commit of the
the global refactoring.