[C++ API] Step 8: Add files for the refactored aten tuner
This PR adds the files implementing the refactored AtenAutotuner API (#345).

This refactoring essentially:
1. makes the tuner future-proof by templating by backend type and search strategy
2. drops the user-exposed load function which retrieves options from a
3. serialized proto cache: the tuner has no business providing this functionality.
If one wants to query the options cache one can do so explicitly
4. drops the owning pointer to a genetic autotuner: an ATenAutotuner is a
Autotuner<Backend, SearchStrategy> that operates on ATen tensors.
5. adds documentation and example of usage

As part of the bigger refactoring, these files are not yet activated
(or even compiled) therefore the changeset cannot be tested independently.
This new implementation will be switched in as part of the last commit of the
the global refactoring.