[C++ API] Add CpuMappingOptions for refactor
This changeset add a few supporting functions that will allow an end-to-end
mock CPU example to compile and run, in a subsequent changeset.
In particular this introduces:
1. extra functions in parameters.cc|h for CPU backend-specific interaction
between search parameters and CpuMappingOptions. These will be used in a
subsequent commit. Note that the mapping options for multiple backends appear
in parameters.cc which is not split. This is because a. options can currently
be used in cross-compilation mode and b. I am punting on non-essential changes
to finish the refactoring;
2. extra methods in CpuMappingOptions to allow it to be used as a template type
in subsequent commits where we hook up the templated autotuner to the examples;
3. a simple CpuMappingOptionsCppPrinter and an operator<< to also allow usage
as a template type.