[C++ API] Move polyhedral mapping types
This commit prepares the polyhedral mapping types for better interaction
with the multi-backend refactoring.
In particular mapping types can be used in cross-compilation situations.
So this commit:
1. moves the mapping types to tc/core/polyhedral and will allow
backend-agnostic portions of the toolchain to fully break the dependency on
cuda-specific files in a subsequent commit;
2. makes mappingSize a backend-specific free function. This breaks the
dependency of mapping types on Grid and Block
3. start using tc/core/polyhedral/mapping_types.h instead of the former cuda
path.